requirements.txt
Django>=4.2
djangorestframework
django-cors-headers
numpy
pandas
scikit-learn
joblib
librosa
soundfile
python-parselmouth
pillow
tensorflow>=2.0   
